Update README.md
Submission:

– A single Java file named Deadlock.java


– Two output files named output01.txt and output02.txt


Introduction


There are a variety of approaches to deal with resource deadlocks. For this assignment you’ll investigate deadlock detection. The approach to deadlock detection that we discussed in class is to model resource allocations and requests with a Resource Allocation Graph (RAG). Deadlock detection is then accomplished by finding a cycle in the graph.


First, you will need to create graph package. The operations (AddNode, AddEdge, RemoveEdge, etc.) will be required. You can start writing this as you become acquainted with the rest of the assignment. Remember that you will need directed edges to be supported by your graph package.


You will definitely want to think about this assignment a lot before you start writing code.